Added code to find and display IP addresses in the input text using regular expressions.
